feat: 농장 시스템 개선 및 7단계 작물 성장 시스템 적용
- 작물 성장 단계 임계값 수정 (4 → 6, 7단계 시스템)
- 작물 아이콘 에셋 추가 (Carrot, Corn, Grape 등)
- 도구 스프라이트 추가 (Axe, Shovel, Watering Can 등)
- PlayerController 도구 사용 애니메이션 개선
- cropConfig 설정 확장
- Hotbar UI 개선

+++ backend/app/models/analysis.py
@@ -117,9 +117,9 @@ class HintIndependence(BaseModel):
 class EloChange(BaseModel):
     """ELO 변화 기록 (단일 토픽)."""
     topic: str
-    before: int
-    after: int
-    change: int
+    before: float
+    after: float
+    change: float
     expected: float
 
 
@@ -128,7 +128,7 @@ class EloHistoryEntry(BaseModel):
     date: str
     topics: List[str] = []
     is_correct: bool = False
-    problem_elo: int = 1000
+    problem_elo: float = 1000
     changes: List[EloChange] = []
